+++ tcp/nc.go
