+++ tcp/scanner.go
